Tekisite/requirements/base.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-filter==23.2
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
drf-social-oauth2==2.1.2
ecdsa==0.18.0
idna==3.4
jwcrypto==1.4.2
Markdown==3.4.3
oauthlib==3.2.2
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.7.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.5.0
urllib3==2.0.2
wrapt==1.15.0
